Replace guessed buttons_held/held with official TPadPort names + types
Per owner directive (our names/types always lose to official ones):
- TPadPort field held -> button (the demo's official field-0 name);
  the u16 type IS the official 'unsigned short', no hedge comment kept.
- Delete the invented buttons_held typedef. GetRealPad now returns the
  official 'long' (s32), GetPad the official 'short' (s16) — both already
  matched in some callers (FUN_800514d8 used s32). Unified ALL GetRealPad
  externs to s32 and ALL GetPad externs to s16(short no), matching the
  PSX.SYM prototypes; every per-file u16/u32 return-type lever turned out
  to reproduce the bytes as the official type (incl. Briefing, whose u32
  lever is now the official s32 — comment updated).
- Eliminate the redundant local PadProcPort/PadShockPort views (they only
  touch act1/act2, same offsets as TPadPort) in favor of TPadPort. ComPad
  keeps ComPadPort — it needs SIGNED x/y for its analog writes; its field-0
  is now 'button' too.
Verified byte-identical.

+++ src/main.exe/StageEndScreen.c
@@ -121,7 +121,7 @@ extern BackGround *FUN_8004f4f8(u_long *tim);
 extern void vfree(void *ptr);
 extern u_long *get_tim_from_archive(u_long *archive, s16 index);
 extern void _PlayMusic(s32 music, s32 mode);
-extern u16 GetRealPad(s32 port);
+extern s32 GetRealPad(s32 port);
 extern void StartDrawing(void);
 extern void DrawBG(BackGround *background);
 extern void FUN_800515b0(GsSPRITE *sprite, s32 value, s32 x, s32 y, s32 mode);
